+++ lib/osm/member.rb
@@ -4,9 +4,9 @@ module Osm
 
     attr_reader :id, :section_id, :type, :first_name, :last_name, :email1, :email2, :email3, :email4, :phone1, :phone2, :phone3, :phone4, :address, :address2, :date_of_birth, :started, :joined_in_years, :parents, :notes, :medical, :religion, :school, :ethnicity, :subs, :grouping_id, :grouping_leader, :joined, :age, :joined_years
     # @!attribute [r] id
-    #   @return [FixNum] the id for the member
+    #   @return [Fixnum] the id for the member
     # @!attribute [r] section_id
-    #   @return [FixNum] the section the member belongs to
+    #   @return [Fixnum] the section the member belongs to
     # @!attribute [r] type
     #   @return [?] ?
     # @!attribute [r] first_name
@@ -38,7 +38,7 @@ module Osm
     # @!attribute [r] started
     #   @return [Date] when the member started Scouting
     # @!attribute [r] joined_in_years
-    #   @return [FixNum] ?
+    #   @return [Fixnum] ?
     # @!attribute [r] parents
     #   @return [String] the member's parent's names
     # @!attribute [r] notes
@@ -54,15 +54,15 @@ module Osm
     # @!attribute [r] subs
     #   @return [String] details about the member's subs
     # @!attribute [r] grouping_id
-    #   @return [FixNum] the grouping within the section that the member belongs to
+    #   @return [Fixnum] the grouping within the section that the member belongs to
     # @!attribute [r] grouping_leader
-    #   @return [FixNum] wether the member is the grouping leader (0=no, 1=seconder/APL, 2=sixer/PL)
+    #   @return [Fixnum] wether the member is the grouping leader (0=no, 1=seconder/APL, 2=sixer/PL)
     # @!attribute [r] joined
     #   @return [Date] when the member joined the section
     # @!attribute [r] age
     #   @return [String] the member's current age (yy/mm)
     # @!attribute [r] joined_years
-    #   @return [FixNum] how many years the member has been in Scouting
+    #   @return [Fixnum] how many years the member has been in Scouting
 
 
     # Initialize a new Member using the hash returned by the API call
@@ -101,13 +101,13 @@ module Osm
     end
 
     # Get the years element of this scout's age
-    # @return [FixNum] the number of years this scout has been alive
+    # @return [Fixnum] the number of years this scout has been alive
     def age_years
       return @age[0..1].to_i
     end
 
     # Get the months element of this scout's age
-    # @return [FixNum] the number of months since this scout's last birthday
+    # @return [Fixnum] the number of months since this scout's last birthday
     def age_months
       return @age[-2..-1].to_i
     end